Starting to create visual structures
This time, we begin creating the visual structures that will be part of our system. We will start by creating two pages: an index page, which will have a search bar for the user to type what they are looking for, and a second page to list the results based on the user's search.