Creating a configuration file and a logging class
This time, we created a file to be responsible for the connection with the database so that it can be invoked whenever we need to list the records stored in the database. We also created functions responsible for managing the records in the database, including functions to create, edit, delete, and delete all records.